requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pandas==2.2.2
Pillow==10.4.0
tqdm==4.66.4
streamlit==1.37.1
openai==1.40.0
# Optional heavy deps for BLIP-2:
# torch
# torchvision
# transformers
